simplify prompts page
